requirements.txt
absl-py==1.2.0
albumentations==1.3.0
cachetools==5.2.0
certifi==2022.9.24
charset-normalizer==2.1.1
colorama==0.4.5
contourpy==1.0.5
cycler==0.11.0
easydict==1.10
et-xmlfile==1.1.0
ever-beta==0.2.3
fonttools==4.37.4
google-auth==2.12.0
google-auth-oauthlib==0.4.6
grpcio==1.49.1
idna==3.4
imageio==2.22.1
imgaug==0.4.0
importlib-metadata==5.0.0
joblib==1.2.0
kiwisolver==1.4.4
Markdown==3.4.1
MarkupSafe==2.1.1
matplotlib==3.6.1
MedPy==0.4.0
mkl-fft==1.3.1
mkl-random==1.2.2
mkl-service==2.4.0
networkx==2.8.7
numpy==1.23.1
oauthlib==3.2.1
opencv-python==4.6.0.66
opencv-python-headless==4.6.0.66
openpyxl==3.0.10
packaging==21.3
pandas==1.5.0
Pillow==9.2.0
pip==22.2.2
prettytable==3.4.1
protobuf==3.19.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2022.4
PyWavelets==1.4.1
PyYAML==6.0
qudida==0.0.4
requests==2.28.1
requests-oauthlib==1.3.1
rsa==4.9
scikit-image==0.19.3
scikit-learn==1.1.2
scipy==1.9.2
setuptools==63.4.1
Shapely==1.8.5.post1
SimpleITK==2.2.1
six==1.16.0
tensorboard==2.10.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
threadpoolctl==3.1.0
tifffile==2022.10.10
torch==1.7.1
torchaudio==0.7.2
torchvision==0.8.2
tqdm==4.64.1
typing_extensions==4.3.0
urllib3==1.26.12
wcwidth==0.2.5
Werkzeug==2.2.2
wheel==0.37.1
wincertstore==0.2
xlwt==1.3.0
zipp==3.9.0
